adding average stats
